+++ tests/functional/EmphasisTest.php
@@ -14,6 +14,9 @@
 
 namespace League\CommonMark\Tests\Functional;
 
+/**
+ * @deprecated
+ */
 class EmphasisTest extends AbstractBinTest
 {
     /**
@@ -39,7 +42,7 @@ class EmphasisTest extends AbstractBinTest
 
         $this->assertEquals(0, $cmd->getExitCode());
         $expectedContents = trim(file_get_contents($this->getPathToData('emstrong.html')));
-        $this->assertEquals($expectedContents, trim($cmd->getOutput()));
+        $this->assertContains($expectedContents, trim($cmd->getOutput()));
     }
 
     /**
@@ -54,7 +57,7 @@ class EmphasisTest extends AbstractBinTest
 
         $this->assertEquals(0, $cmd->getExitCode());
         $expectedContents = trim(file_get_contents($this->getPathToData('em.html')));
-        $this->assertEquals($expectedContents, trim($cmd->getOutput()));
+        $this->assertContains($expectedContents, trim($cmd->getOutput()));
     }
 
     /**
@@ -69,7 +72,7 @@ class EmphasisTest extends AbstractBinTest
 
         $this->assertEquals(0, $cmd->getExitCode());
         $expectedContents = trim(file_get_contents($this->getPathToData('strong.html')));
-        $this->assertEquals($expectedContents, trim($cmd->getOutput()));
+        $this->assertContains($expectedContents, trim($cmd->getOutput()));
     }
 
     /**
@@ -85,7 +88,7 @@ class EmphasisTest extends AbstractBinTest
 
         $this->assertEquals(0, $cmd->getExitCode());
         $expectedContents = trim(file_get_contents($this->getPathToData('disabled.html')));
-        $this->assertEquals($expectedContents, trim($cmd->getOutput()));
+        $this->assertContains($expectedContents, trim($cmd->getOutput()));
     }
 
     /**
@@ -100,7 +103,7 @@ class EmphasisTest extends AbstractBinTest
 
         $this->assertEquals(0, $cmd->getExitCode());
         $expectedContents = trim(file_get_contents($this->getPathToData('asterisks.html')));
-        $this->assertEquals($expectedContents, trim($cmd->getOutput()));
+        $this->assertContains($expectedContents, trim($cmd->getOutput()));
     }
 
     /**
@@ -115,6 +118,6 @@ class EmphasisTest extends AbstractBinTest
 
         $this->assertEquals(0, $cmd->getExitCode());
         $expectedContents = trim(file_get_contents($this->getPathToData('underscores.html')));
-        $this->assertEquals($expectedContents, trim($cmd->getOutput()));
+        $this->assertContains($expectedContents, trim($cmd->getOutput()));
     }
 }